requirements.txt
# GPU-Insight 依赖

# 核心
pyyaml>=6.0
python-dotenv>=1.0

# LLM API
anthropic>=0.40.0
openai>=1.50.0

# 爬虫
playwright>=1.40
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0

# 数据处理
opencc-python-reimplemented>=0.1.7

# Web 界面
fastapi>=0.115
uvicorn>=0.30
jinja2>=3.1

# 测试
pytest>=8.0
pytest-asyncio>=0.24
